third_party/WebKit/LayoutTests/accessibility/accessibility-object-detached-expected.txt

This test makes sure that AccessibilityObjects are detached when the node they point to is detached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expectedButtonRole != expectedDetachedRole is true
PASS successfullyParsed is true

TEST COMPLETE
